Fixing issue #75
Turns out there was a tag missing that needed to be added to make the Interfusions show up in preview like that.  Getting rid of the human-only interfusions seems to be more trouble than it's worth, but something could probably be done if we override those recipes directly and block them from working on Drauven.  Seems overkill and needlessly incompatible to add that on as well though.